Improving the position of the highcharts elements.
