+++ verticapy/plotting/_highcharts/heatmap.py
@@ -20,6 +20,8 @@ from verticapy._config.colors import get_colors
 from verticapy._utils._sql._sys import _executeSQL
 from verticapy.connection import current_cursor
 
+from verticapy.plotting._highcharts.utils import data_to_columns
+
 
 def heatmap(
     query: str = "",
@@ -27,9 +29,11 @@ def heatmap(
     options: dict = {},
     width: int = 600,
     height: int = 400,
-):
-    from verticapy.plotting._highcharts.base import data_to_columns
-
+) -> Highchart:
+    """
+    Draws a heatmap using the High Chart API 
+    and the input SQL query.
+    """
     chart = Highchart(width=width, height=height)
     default_options = {
         "chart": {
